2025_QRgenerate/requirements.txt
streamlit==1.32.0
pandas==2.2.1
qrcode[pil]==7.4.2
reportlab==4.1.0
google-cloud-storage==2.16.0
gcsfs==2024.3.1
